Issue with python command
Due to Python 2.7 being updated to Python 3 on our VMs, it seems that the python command is saved as python3. It means that when following the workbook and attempting to run with 'python', we get a 'command not found' error. Tried looking for instances where it occurs, mostly an issue in chapter 4.1